requirements.txt
pyrtlsdr
scapy
git+https://github.com/pybluez/pybluez.git#egg=pybluez
numpy